# apps/server/test/unit/common/__snapshots__/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/menu/current": {
      "get": {
        "operationId": "MenuController_current",
        "summary": "获取当前发布菜单",
        "parameters": [
          {
            "name": "authorizerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/menu/draft": {
      "get": {
        "operationId": "MenuController_draft",
        "summary": "获取菜单草稿",
        "parameters": [
          {
            "name": "authorizerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/menu": {
      "post": {
        "operationId": "MenuController_save",
        "summary": "保存菜单草稿",
        "parameters": [
          {
            "name": "authorizerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/menu/publish": {
      "post": {
        "operationId": "MenuController_publish",
        "summary": "发布菜单到微信",
        "parameters": [
          {
            "name": "authorizerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/menu/history": {
      "get": {
        "operationId": "MenuController_history",
        "summary": "发布历史",
        "parameters": [
          {
            "name": "authorizerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/menu/templates": {
      "get": {
        "operationId": "MenuController_templates",
        "summary": "菜单模板列表",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "MenuController_createTemplate",
        "summary": "保存为模板",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/menu/templates/{templateId}": {
      "delete": {
        "operationId": "MenuController_deleteTemplate",
        "summary": "删除模板",
        "parameters": [
          {
            "name": "templateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/menu/templates/{templateId}/apply": {
      "post": {
        "operationId": "MenuController_applyTemplate",
        "summary": "应用模板（替换当前草稿）",
        "parameters": [
          {
            "name": "authorizerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "templateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/campaigns": {
      "get": {
        "operationId": "CampaignController_list",
        "summary": "活动列表",
        "parameters": [
          {
            "name": "authorizerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "营销活动"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "CampaignController_create",
        "summary": "创建活动",
        "parameters": [
          {
            "name": "authorizerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "营销活动"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/campaigns/{id}": {
      "get": {
        "operationId": "CampaignController_detail",
        "summary": "活动详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "营销活动"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "CampaignController_update",
        "summary": "编辑活动",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "营销活动"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CampaignController_delete",
        "summary": "删除活动",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "营销活动"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/campaigns/{id}/{action}": {
      "post": {
        "operationId": "CampaignController_changeStatus",
        "summary": "活动状态操作 (start/pause/end)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "营销活动"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/campaigns/qrcodes/list": {
      "get": {
        "operationId": "CampaignController_qrList",
        "summary": "渠道码列表",
        "parameters": [
          {
            "name": "authorizerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "campaignId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "营销活动"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/campaigns/qrcodes": {
      "post": {
        "operationId": "CampaignController_createQr",
        "summary": "创建渠道二维码",
        "parameters": [
          {
            "name": "authorizerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "营销活动"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/campaigns/qrcodes/{id}": {
      "delete": {
        "operationId": "CampaignController_deleteQr",
        "summary": "删除渠道码",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "营销活动"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "WXGZH",
    "description": "",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
...
}